feat(mvp): HANDOFF-03 T1-T9 complete, T3/T4/T5 in progress
Implements the demo-grade MVP backend per
docs/handoffs/HANDOFF-03-mvp-backend-tasks.md (v0.5).

Complete:
- T1  money -> integer VND. All money columns Numeric(20,0) across lending,
      ledger, contracts, market and payments. Distribution splitter rewritten
      as pure-integer Hamilton apportionment. Migration ca8a6db6183e.
- T2  ARQ worker scaffold (app/scheduler/), Redis config, compose services.
- T15 admin-gated KYC document approval (GAP-1) — unblocks listing.
- T6  facility data model: Facility, ScheduleVersion, ScheduledPayment,
      InboundTransfer, StateTransition. Append-only, DB-trigger enforced.
      Migration 9142fb4c8540.
- T9  table-driven facility state machine with guards and INV-4.
- T8  Vietnamese business-day calendar incl. Tet (2026 confirmed,
      2027 provisional).
- T7  origination and schedule generation. Reproduces repayment spec
      §12.1 exactly: I=12,000,000; T0=312,000,000; N0=88; A0=3,545,454;
      remainder 48; final 3,545,502; N_bs=118.

In progress:
- T3  sector_reference.py loader; the 72.50/"B" mock is replaced with real
      grade() calls. Gate 3 enabled at threshold 13 (T0a, Loc).
- T4  score-run persistence and replay; bank_rate_config table with
      effective dates (T0b, Loc). Migration 298cd240e4f1.
- T5  working_days_per_month 21 -> 22; params_version wb-v1-20260917.
      working_days_per_year stays 252 — a distinct revenue-annualisation
      convention, not a schedule count.

Also commits the repayment mechanism spec v1.3 itself, which had never been
in the repo — its absence forced the backstop formula and state-transition
table to be reconstructed rather than read.

Not started: T10, T14, T16, T20, T21.

Tests are written but have NOT been run against a live database.

+++ app/contracts/schemas.py
@@ -1,7 +1,6 @@
 from uuid import UUID
 
 from pydantic import BaseModel
-from decimal import Decimal
 
 
 class ContractCreate(BaseModel):
@@ -15,7 +14,7 @@ class ContractOut(BaseModel):
     application_id: UUID | None
     score_run_id: UUID | None
     status: str
-    target_amount: Decimal
-    funded_amount: Decimal
+    target_amount: int
+    funded_amount: int
 
     model_config = {"from_attributes": True}
